feat: make delete drive part of settings modal
