--- src/connect/components/delete-drive-modal/index.ts
@@ -1,1 +1,0 @@
-export * from './delete-drive-modal';
